# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * King SMM Panel API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface User {
  id: string;
  /** @nullable */
  email: string | null;
  /** @nullable */
  firstName: string | null;
  /** @nullable */
  lastName: string | null;
  /** @nullable */
  profileImageUrl: string | null;
  createdAt: string;
}

export interface Service {
  id: number;
  name: string;
  category: string;
  description: string;
  pricePerThousand: number;
  minQuantity: number;
  maxQuantity: number;
  platform: string;
  isActive: boolean;
}

export interface Order {
  id: number;
  serviceId: number;
  serviceName: string;
  platform: string;
  link: string;
  quantity: number;
  charge: number;
  status: string;
  createdAt: string;
}

export interface OrdersPage {
  orders: Order[];
  total: number;
  page: number;
  totalPages: number;
}

export interface OrderInput {
  serviceId: number;
  link: string;
  quantity: number;
}

export interface Wallet {
  balance: number;
  totalSpent: number;
  totalAdded: number;
}

export interface TopupInput {
  amount: number;
  paymentMethod: string;
  transactionId: string;
}

export interface TopupRequest {
  id: number;
  amount: number;
  paymentMethod: string;
  transactionId: string;
  status: string;
  createdAt: string;
}

export interface DashboardStats {
  balance: number;
  totalOrders: number;
  completedOrders: number;
  pendingOrders: number;
  totalSpent: number;
  recentOrders: Order[];
}

export interface AdminStats {
  totalUsers: number;
  totalBalance: number;
  totalOrders: number;
  pendingTopups: number;
}

export interface AdminUser {
  id: string;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  firstName?: string | null;
  /** @nullable */
  lastName?: string | null;
  /** @nullable */
  profileImageUrl?: string | null;
  balance: number;
  totalOrders: number;
  createdAt: string;
}

export interface AdminOrder {
  id: number;
  userId: string;
  /** @nullable */
  userEmail?: string | null;
  serviceId: number;
  serviceName: string;
  platform: string;
  link: string;
  quantity: number;
  charge: number;
  status: string;
  createdAt: string;
}

export interface AdminOrdersPage {
  orders: AdminOrder[];
  total: number;
  page: number;
  totalPages: number;
}

export interface AdminTopup {
  id: number;
  userId: string;
  /** @nullable */
  userEmail?: string | null;
  amount: number;
  paymentMethod: string;
  transactionId: string;
  status: string;
  createdAt: string;
}

export interface AdminTopupsPage {
  topups: AdminTopup[];
  total: number;
  page: number;
  totalPages: number;
}

export interface BalanceAdjustInput {
  balance: number;
}

export interface TopupStatusInput {
  status: string;
}

export interface PaymentSettings {
  upiId: string;
  qrUrl: string;
}

export interface ApiSettings {
  apiUrl: string;
  apiKey: string;
  isEnabled: boolean;
}

export interface ServiceInput {
  name: string;
  category: string;
  platform: string;
  description: string;
  pricePerThousand: number;
  minQuantity: number;
  maxQuantity: number;
}

export type ListServicesParams = {
category?: string;
};

export type ListOrdersParams = {
status?: string;
search?: string;
page?: number;
limit?: number;
};

export type ListAdminOrdersParams = {
page?: number;
limit?: number;
status?: string;
};

export type ListAdminTopupsParams = {
page?: number;
limit?: number;
status?: string;
};

export type DeleteService200 = {
  success: boolean;
};