# lib/api-zod/src/generated/types/service.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * King SMM Panel API
 * OpenAPI spec version: 0.1.0
 */

export interface Service {
  id: number;
  name: string;
  category: string;
  description: string;
  pricePerThousand: number;
  minQuantity: number;
  maxQuantity: number;
  platform: string;
  isActive: boolean;
}